feat(backup): add JSON import with id conflict resolution and tagId remapping via Room transaction

+++ app/src/main/java/com/app/noteapp/domain/backup_model/TagBackup.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 data class TagBackupDto(
     val id: Long,
     val name: String,
-    val colorArgb: Long,
+    val colorArgb: Int,
 )
